socialai/WebApp-API-develop/notifications/requirements.txt
Flask-SqlAlchemy==2.4.1
boto3==1.10.28
Python-dotenv==0.10.3